In RFST_IDLE, block RF_DISCOVER_RSP callback only if DISABLING
If an RF_DEACTIVATE_CMD is sent just after an RF_DISCOVER_CMD
when in RFST_IDLE, the command should be sent and a callback done
to the JNI to unblock any waiting process/mutex. The only case when
this callback is not needed is when the stack is disabling.

Bug: 374687433
Test: manual
Change-Id: Ic57e367cdc674ddf5ced045b1b5e4e0a79e7a45a